requirements.txt
pathspec
gitpython
